refactor(runtime): add slot receiver to app services
